+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+-   Complete refactoring with initial support for GUI patching (SJSON + Lua).
+
 ## [v0.1.0] - 2021-05-24
 
 ### Added